requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.17.0
openpyxl==3.1.2
pyodbc==4.0.30
datetime==4.3
requests==2.26.0
pdfkit==0.6.1
Jinja2==3.1.2
azure-storage-blob==12.17.0
pandas==1.3.3
xlsxwriter==1.4.3
pdfservices-sdk==2.3.0
Flask==2.0.1
Werkzeug==2.0.1
openai